+++ src/main/model/querys_db.ts
@@ -234,19 +234,19 @@ export class Queries {
   SQL_DEPENDENCIES_INSERT =
     'INSERT OR IGNORE INTO dependencies (fileId, purl, version, url, scope, licenses, component,originalVersion,originalLicense) VALUES (?,?,?,?,?,?,?,?,?);';
 
-  SQL_GET_ALL_DEPENDENCIES = `SELECT f.path, d.dependencyId, d.component AS componentName, d.purl, d.version, d.licenses, d.component, d.scope, i.id AS inventory,cv.id AS compid,d.rejectedAt,(CASE WHEN i.id IS NOT NULL AND d.rejectedAt IS NULL THEN '${FileStatusType.IDENTIFIED}' WHEN d.rejectedAt IS NOT NULL THEN '${FileStatusType.ORIGINAL}' ELSE '${FileStatusType.PENDING}' END) AS status,(CASE WHEN d.purl IS NOT NULL AND d.version IS NOT NULL AND licenses IS NOT NULL THEN true ELSE false END) AS valid,d.originalVersion,d.originalLicense, d.fileId FROM dependencies d
+  SQL_GET_ALL_DEPENDENCIES = `SELECT f.path, d.dependencyId, d.component AS componentName, d.purl, d.version, d.licenses, d.component, d.scope, i.id AS inventory,cv.id AS compid,d.rejectedAt,(CASE WHEN i.id IS NOT NULL AND d.rejectedAt IS NULL THEN '${FileStatusType.IDENTIFIED}' WHEN d.rejectedAt IS NOT NULL THEN '${FileStatusType.ORIGINAL}' ELSE '${FileStatusType.PENDING}' END) AS status,(CASE WHEN d.purl IS NOT NULL AND d.version IS NOT NULL AND licenses IS NOT NULL THEN true ELSE false END) AS valid,d.originalVersion,d.originalLicense, d.fileId, d.url FROM dependencies d
   INNER JOIN files f ON f.fileId =  d.fileId
   LEFT JOIN component_versions cv ON cv.purl= d.purl AND cv.version = d.version
   LEFT JOIN inventories i ON cv.id = i.cvid  AND instr(d.licenses, i.spdxid)>0 #FILTER AND (i.source = 'declared' OR i.source IS NULL);`;
 
-  SQL_ALL_IDENTIFIED_DEPENDENCIES = `SELECT f.path as file ,d.component,d.purl,d.version,d.licenses
+  SQL_ALL_IDENTIFIED_DEPENDENCIES = `SELECT f.path as file ,d.component,d.purl,d.version,d.licenses, d.url
   FROM dependencies d
   INNER JOIN files f ON f.fileId =  d.fileId
-  INNER JOIN inventories i ON cv.id = i.cvid
   INNER JOIN component_versions cv ON cv.purl= d.purl AND cv.version = d.version
+  INNER JOIN inventories i ON cv.id = i.cvid
   GROUP BY f.path, d.component, d.version, d.purl;`;
 
-  SQL_ALL_DETECTED_DEPENDENCIES = `SELECT f.path as file ,d.component,d.purl,d.version,d.originalLicense as licenses FROM dependencies d
+  SQL_ALL_DETECTED_DEPENDENCIES = `SELECT f.path as file ,d.component,d.purl,d.version,d.originalLicense as licenses, d.url FROM dependencies d
   INNER JOIN files f ON f.fileId =  d.fileId;`;
 
   SQL_GET_ALL_FILES = `SELECT f.path, r.md5_file, f.fileId AS id,f.identified,f.ignored,(CASE WHEN f.identified=0 AND f.ignored=0 THEN 1 ELSE 0 END) AS pending, f.type FROM files f