+++ src/main/workspace/tree/Folder.ts
@@ -239,16 +239,14 @@ export default class Folder extends Node {
   public summarize(root: string, summary: any): any {
     if (this.getAction() === 'filter') {
       this.setClassName('filter-item');
-    } else {
-      this.children.forEach((child) => {
+    }
+    this.children.forEach((child) => {
         const aux = child.summarize(root, summary);
         summary.total = aux.total;
         summary.include = aux.include;
         summary.filter = aux.filter;
         summary.files = aux.files;
       });
-    }
-
     return summary;
   }