+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Upcoming changes ...
+### Fixed
+- Fixed errors when closing a project while async operations are in-flight
 
 ## [1.29.0] - 2025-12-02
 ### Added